modules/person-grpc/requirements.txt
SQLAlchemy==1.3.19
psycopg2-binary==2.8.5
GeoAlchemy2==0.8.4
grpcio==1.43.0
protobuf==3.19.3
shapely==1.7.0